M02A01 – EX007

+++ Modulo_02_Operadores/M02A01–EX007.c
@@ -8,7 +8,7 @@ void main(){
     printf("<<< EX007 - Dobro e Terça parte >>> \n\n");
     printf("Digite um número: ");
     scanf("%d", &num);
-    printf("Analisando o número %d, seu dobro é %d e sua terça parte é %d. \n", num, (num*2), (num/3));
+    printf("Analisando o número %d, seu dobro é %d e sua terça parte é %.2f. \n", num, (num*2), ((float)num/3));
 }
 
 /* <<< EX007 - Dobro e Terça parte >>>
